update fedex.php to fix undefined errors and remove personal data
Several changes.
Fix undefined error when response timestamp is empty when FedEx doesn't give a time estimate.
Add Canada province code as required to return a response for Canada.
Fix incorrect province error for other countries by making it empty. Certain countries require specific 2 letter province codes that opencart does not have. So far the rates seem to be correct without the province.
Add 12 hours to the shipping date when the current time is after 12:00. Because even if it's 23:59 and far past the cutoff time, the rate response still thinks you want to ship it today.
Change the cost returned to TotalNetFedExCharge, which is the real net charge without tax. TotalNetCharge includes sales tax based on the shipper's address. So if your opencart is set to prices exclusive of tax, then using TotalNetCharge will double tax the shipping rate.